Реализация кнопки Show more (правки)
